+++ nym-vpn-app/src/components/index.ts
@@ -1,1 +1,2 @@
 export * from './back-navigation-confirmation-dialog';
+export * from './privy';