+++ nym-vpn-apple/ServicesIOS/Sources/ErrorHandler/VPNErrorReason.swift
@@ -5,7 +5,6 @@ import Theme
 
 public enum VPNErrorReason: LocalizedError {
     case initialization(details: String)
-    case createLogFile(details: String)
     case internalError(details: String)
     case storage(details: String)
     case networkConnectionError(details: String)
@@ -33,6 +32,9 @@ public enum VPNErrorReason: LocalizedError {
     case nyxdConnectionFailure(details: String)
     case nyxdQueryFailure(details: String)
     case invalidSecret(details: String)
+    case initLogs(details: String)
+    case deeplinkError(details: String)
+
     case unkownTunnelState
 
     private static let somethingWentWrong = "generalNymError.somethingWentWrong".localizedString
@@ -46,8 +48,6 @@ public enum VPNErrorReason: LocalizedError {
         switch vpnError {
         case let .Initialization(details: details):
             self = .initialization(details: details)
-        case let .CreateLogFile(details: details):
-            self = .createLogFile(details: details)
         case let .InternalError(details: details):
             self = .internalError(details: details)
         case let .Storage(details: details):
@@ -222,6 +222,10 @@ public enum VPNErrorReason: LocalizedError {
             self = .nyxdQueryFailure(details: details)
         case let .InvalidSecret(details: details):
             self = .invalidSecret(details: details)
+        case let .InitLogs(details: details):
+            self = .initLogs(details: details)
+        case let .DeeplinkError(details: details):
+            self = .deeplinkError(details: details)
         }
     }
 
@@ -237,8 +241,6 @@ public enum VPNErrorReason: LocalizedError {
         switch errorReason {
         case .initialization:
             self = .initialization(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
-        case .createLogFile:
-            self = .createLogFile(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
         case .internalError:
             self = .internalError(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
         case .storage:
@@ -297,6 +299,10 @@ public enum VPNErrorReason: LocalizedError {
             self = .nyxdQueryFailure(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
         case .invalidSecret:
             self = .invalidSecret(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
+        case .initLogs:
+            self = .initLogs(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
+        case .deeplinkError:
+            self = .deeplinkError(details: nsError.userInfo["details"] as? String ?? Self.somethingWentWrong)
         }
     }
 
@@ -326,8 +332,6 @@ extension VPNErrorReason {
     var description: String {
         switch self {
         case let .internalError(details):
-            details
-        case let .createLogFile(details):
             details
         case let .storage(details):
             details
@@ -385,6 +389,10 @@ extension VPNErrorReason {
             details
         case let .invalidSecret(details: details):
             details
+        case let .initLogs(details: details):
+            details
+        case let .deeplinkError(details: details):
+            details
         }
     }
 }
@@ -398,7 +406,6 @@ extension VPNErrorReason: Equatable {
 /// The VPNErrorReasonCode mirrors the error codes as raw integers and can be constructed from a VPNErrorReason.
 enum VPNErrorReasonCode: Int, RawRepresentable {
     case initialization
-    case createLogFile
     case internalError
     case storage
     case networkConnectionError
@@ -427,13 +434,13 @@ enum VPNErrorReasonCode: Int, RawRepresentable {
     case nyxdConnectionFailure
     case nyxdQueryFailure
     case invalidSecret
+    case initLogs
+    case deeplinkError
 
     init?(vpnErrorReason: VPNErrorReason) {
         switch vpnErrorReason {
         case .initialization:
             self = .initialization
-        case .createLogFile:
-            self = .createLogFile
         case .internalError:
             self = .internalError
         case .storage:
@@ -490,6 +497,10 @@ enum VPNErrorReasonCode: Int, RawRepresentable {
             self = .nyxdQueryFailure
         case .invalidSecret:
             self = .invalidSecret
+        case .initLogs:
+            self = .initLogs
+        case .deeplinkError:
+            self = .deeplinkError
         }
     }
 }