+++ nym-vpn-core/crates/nym-vpn-account-controller/src/commands/common_handler.rs
@@ -11,6 +11,7 @@ use nym_vpn_lib_types::{AccountCommandError, VpnAccountSummary};
 use crate::{
     AvailableTicketbooks, SharedAccountState,
     commands::{ReturnSender, dispatch::CommonCommand},
+    deeplink::CreateDeeplinkParams,
     storage::AccountStorageOp,
 };
 use nym_vpn_store::account::StorableAccount;
@@ -47,6 +48,11 @@ pub(crate) async fn handle_common_command<C: ConnectivityMonitor>(
         CommonCommand::GetAccountSummary(result_tx) => {
             result_tx.send(handle_get_account_summary(shared_state).await);
         }
+        CommonCommand::GetDeeplink(result_tx, params) => {
+            result_tx.send(handle_get_deeplink(shared_state, params).await)
+        }
+        CommonCommand::DeriveDeeplinkMnemonic(result_tx, deeplink_callback_url) => result_tx
+            .send(handle_derive_deeplink_mnemonic(shared_state, deeplink_callback_url).await),
     };
 }
 
@@ -176,4 +182,39 @@ pub(crate) async fn handle_get_account_summary<C: ConnectivityMonitor>(
     shared_state: &mut SharedAccountState<C>,
 ) -> Result<Option<VpnAccountSummary>, AccountCommandError> {
     Ok(shared_state.vpn_account_summary.clone())
+}
+
+pub(crate) async fn handle_get_deeplink<C: ConnectivityMonitor>(
+    shared_state: &mut SharedAccountState<C>,
+    params: CreateDeeplinkParams,
+) -> Result<String, AccountCommandError> {
+    // Create a new Deeplink for this request
+    let deeplink = shared_state
+        .deeplinks
+        .create_deeplink(&params)
+        .map_err(|e| AccountCommandError::DeeplinkError(e.to_string()))?;
+
+    // Create the deeplink URL
+    let url = deeplink.create_url(&params.base_url);
+
+    // Housekeeping
+    shared_state.deeplinks.remove_expired();
+
+    Ok(url.to_string())
+}
+
+pub(crate) async fn handle_derive_deeplink_mnemonic<C: ConnectivityMonitor>(
+    shared_state: &mut SharedAccountState<C>,
+    deeplink_callback_url: String,
+) -> Result<bip39::Mnemonic, AccountCommandError> {
+    // Derive the mnemonic from the provided deeplink URL
+    let mnemonic = shared_state
+        .deeplinks
+        .derive_mnemonic(&deeplink_callback_url)
+        .map_err(|e| AccountCommandError::DeeplinkError(e.to_string()))?;
+
+    // Housekeeping
+    shared_state.deeplinks.remove_expired();
+
+    Ok(mnemonic)
 }