+++ nym-vpn-core/crates/nym-vpn-account-controller/src/commands/dispatch.rs
@@ -1,7 +1,7 @@
 // Copyright 2025 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
-use crate::AvailableTicketbooks;
+use crate::{AvailableTicketbooks, deeplink::CreateDeeplinkParams};
 use nym_validator_client::nyxd::Coin;
 use nym_vpn_api_client::{
     ResolverOverrides,
@@ -84,6 +84,10 @@ impl AccountCommand {
                     return_sender.send(Err(error))
                 }
                 CommonCommand::GetAccountSummary(return_sender) => return_sender.send(Err(error)),
+                CommonCommand::GetDeeplink(return_sender, _) => return_sender.send(Err(error)),
+                CommonCommand::DeriveDeeplinkMnemonic(return_sender, _) => {
+                    return_sender.send(Err(error))
+                }
             },
             AccountCommand::UpgradeMode(upgrade_mode_command) => match upgrade_mode_command {
                 UpgradeModeCommand::GetUpgradeModeEnabled(return_sender) => {
@@ -129,6 +133,15 @@ pub enum CommonCommand {
 
     /// Returns the VPN account summary if the account is logged-in
     GetAccountSummary(ReturnSender<Option<VpnAccountSummary>, AccountCommandError>),
+
+    /// Return the deeplink URL for the specfied deeplink kind and name
+    GetDeeplink(
+        ReturnSender<String, AccountCommandError>,
+        CreateDeeplinkParams,
+    ),
+
+    /// Derive the mnemonic from the deeplink callback URL
+    DeriveDeeplinkMnemonic(ReturnSender<bip39::Mnemonic, AccountCommandError>, String),
 }
 
 /// Commands relating to the upgrade mode