+++ nym-vpn-app/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Privy social login
+- Lewes Protocol switch in DEV settings
+
 ### Fixed
 
 - UI styling